+++ ui/src/commonMain/composeResources/values/strings.xml
@@ -3,6 +3,7 @@
     <string name="navigation_bar_navigate">Navigate</string>
     <string name="navigation_bar_favourites">Favourites</string>
     <string name="navigation_bar_about">About</string>
+    <string name="navigation_bar_more">More</string>
 
     <string name="browse_option_see_all_service_alerts">See all service updates</string>
     <string name="browse_option_upcoming_routes">Upcoming vehicles to %1$s</string>
